Changes to dir
