--- olx_jobseeker/olx_jobseeker/items.py
@@ -1,18 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Define here the models for your scraped items
-#
-# See documentation in:
-# https://docs.scrapy.org/en/latest/topics/items.html
-
-import scrapy
-
-
-class OlxJobseekerItem(scrapy.Item):
-    # define the fields for your item here like:
-    title = scrapy.Field()
-    description = scrapy.Field()
-    link = scrapy.Field() 
-    salary_from = scrapy.Field()
-    salary_to = scrapy.Field()
-
