--- olx_jobseeker/utils/cleaner.py
@@ -1,9 +1,0 @@
-import re
-
-
-def clean_data(text):
-    text = re.sub('<[^>]*>', '', text)
-    text = re.sub('[0-9]*', '', text)
-    text = re.sub(' +', ' ', text)
-    text = re.sub('(\n|\r)+', '', text)
-    return text
